perf(runner): lazy-load dynamic tool runtime
